requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
pyparsing==2.4.0
python-cassandra==0.6.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
thrift==0.11.0
